+++ runtime/networkdriver/lxc/driver.go
@@ -3,10 +3,10 @@ package lxc
 import (
 	"fmt"
 	"github.com/dotcloud/docker/engine"
-	"github.com/dotcloud/docker/networkdriver"
-	"github.com/dotcloud/docker/networkdriver/ipallocator"
-	"github.com/dotcloud/docker/networkdriver/portallocator"
-	"github.com/dotcloud/docker/networkdriver/portmapper"
+	"github.com/dotcloud/docker/runtime/networkdriver"
+	"github.com/dotcloud/docker/runtime/networkdriver/ipallocator"
+	"github.com/dotcloud/docker/runtime/networkdriver/portallocator"
+	"github.com/dotcloud/docker/runtime/networkdriver/portmapper"
 	"github.com/dotcloud/docker/pkg/iptables"
 	"github.com/dotcloud/docker/pkg/netlink"
 	"github.com/dotcloud/docker/utils"
