+++ runtime/runtime.go
@@ -17,8 +17,8 @@ import (
 	_ "github.com/dotcloud/docker/graphdriver/devmapper"
 	_ "github.com/dotcloud/docker/graphdriver/vfs"
 	"github.com/dotcloud/docker/image"
-	_ "github.com/dotcloud/docker/networkdriver/lxc"
-	"github.com/dotcloud/docker/networkdriver/portallocator"
+	_ "github.com/dotcloud/docker/runtime/networkdriver/lxc"
+	"github.com/dotcloud/docker/runtime/networkdriver/portallocator"
 	"github.com/dotcloud/docker/pkg/graphdb"
 	"github.com/dotcloud/docker/pkg/sysinfo"
 	"github.com/dotcloud/docker/runconfig"
